requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
python-dotenv==1.0.1
Pillow==10.2.0
django-cors-headers==4.3.1